+++ R/step_spline_natural.R
@@ -1,0 +1,4 @@
+#' @export
+orbital.step_spline_natural <- function(x, all_vars, ...) {
+  spline_helper(x, all_vars, splines2::naturalSpline)
+}